requirements.txt
# =============================================================================
# AI Algo Trading System — Python Dependencies
# =============================================================================

# --- Config & Environment ---
pyyaml>=6.0
python-dotenv>=1.0.0
tzdata>=2024.1

# --- Logging ---
structlog>=24.1.0

# --- Database ---
# sqlite3 is stdlib, no external dependency needed

# --- Brokers ---
alpaca-py>=0.28.0
kiteconnect>=5.0.0

# --- Market Data ---
yfinance>=0.2.36

# --- AI / LLM ---
anthropic>=0.39.0
langgraph>=0.2.0
langchain-core>=0.3.0

# --- Technical Analysis ---
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0

# --- Web Dashboard ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0

# --- Notifications ---
twilio>=9.0.0

# --- News & Web ---
feedparser>=6.0.0
aiohttp>=3.10.0

# --- Security ---
cryptography>=43.0.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# --- Scheduling ---
apscheduler>=3.10.0
